[SDK] sync changes according to performance benchmarks (#297)
* sync SDK changes according to performance benchmarks

* fix end-of-file lint

* fix clang-format issue

* fix clang-format by adding 'clang-format off'

* remove useless casts

* remove 'data' argument of 'operator()'

* change 'Tensor2Img' to 'TensorToImg' according to spec

* correct tensor's name according spec

+++ csrc/apis/c/common.h
@@ -54,10 +54,10 @@ typedef struct mm_mat_t {
 } mm_mat_t;
 
 typedef struct mm_rect_t {
-  int left;
-  int top;
-  int right;
-  int bottom;
+  float left;
+  float top;
+  float right;
+  float bottom;
 } mm_rect_t;
 
 typedef struct mm_pointi_t {
